feat: create new profile

+++ src/main/java/facebook/app/services/UserService.java
@@ -78,11 +78,11 @@ public class UserService {
         return userDAO.getUserByID(userID);
     }
 
-    public long getCurrentUserId() throws UserIOException {
+    public int getCurrentUserId() throws UserIOException {
         List<User> userList = userDAO.readUsers();
         for (User user : userList) {
             if (user.isLoggedIn()) {
-                return user.getUserId();
+                return (int) user.getUserId();
             }
         }
         return -1; // Return -1 if no user is currently logged in
